Added java doc
