+++ src/main/java/app/Main.java
@@ -1,6 +1,6 @@
 package app;
 
-import view.GUI.MainPanel;
+import view.MainPanel;
 
 import javax.swing.*;
 import java.awt.*;
