--- src/main/java/view/GUI/SearchPanel.java
@@ -1,15 +1,0 @@
-package view.GUI;
-
-import javax.swing.*;
-
-public class SearchPanel extends JPanel{
-
-    private JButton filterButton;
-    private JButton searchButton;
-    private JTextField searchField;
-    private JPanel SearchPanel;
-
-    public SearchPanel() {
-        setSize(400, 1000);
-    }
-}
