+++ src/main/java/view/TimetablePanel.form
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<form xmlns="http://www.intellij.com/uidesigner/form/" version="1" bind-to-class="view.GUI.TimetablePanel">
+<form xmlns="http://www.intellij.com/uidesigner/form/" version="1" bind-to-class="view.TimetablePanel">
   <grid id="27dc6" binding="TimetablePanel" layout-manager="GridLayoutManager" row-count="3" column-count="8" same-size-horizontally="false" same-size-vertically="false" hgap="-1" vgap="-1">
     <margin top="0" left="0" bottom="0" right="0"/>
     <constraints>
